backend/requirements.txt
firebase-admin
fastapi
uvicorn
python-dotenv
pydantic
deepgram-sdk
websockets
google-generativeai
requests